Fix bool streaming in DBG statement

+++ magda/daw/audio/AudioBridge.cpp
@@ -113,7 +113,7 @@ void AudioBridge::trackPropertyChanged(int trackId) {
                                 // Found a MIDI input routed to this track - sync record armed state
                                 inputDeviceInstance->setRecordingEnabled(track->itemID,
                                                                          trackInfo->recordArmed);
-                                DBG("Synced recordArmed=" << trackInfo->recordArmed
+                                DBG("Synced recordArmed=" << (int)trackInfo->recordArmed
                                                           << " to MIDI input '"
                                                           << inputDeviceInstance->owner.getName()
                                                           << "' for track " << trackId);
